fix: 빌드 전 에러사항 수정
