+++ src/app/(providers)/(root)/products/[id]/_components/ButtonsWrapper.tsx
@@ -28,7 +28,8 @@ function ButtonsWrapper({
     const cartItemData = {
       cart_product_id: productId,
       count: count,
-      cart_user_id: user.id
+      cart_user_id: user.id,
+      is_checked: true
     };
     const result = await createCartItem(cartItemData);
